+++ src/main/java/com/besson/endfield/datagen/ModBlockTagsProvider.java
@@ -25,7 +25,17 @@ public class ModBlockTagsProvider extends FabricTagProvider.BlockTagProvider {
                 .add(ModBlocks.IRON_MINERAL_VEIN_BLOCK)
                 .add(ModBlocks.LAPIS_MINERAL_VEIN_BLOCK)
                 .add(ModBlocks.ORIGINIUM_MINERAL_VEIN_BLOCK)
-                .add(ModBlocks.REDSTONE_MINERAL_VEIN_BLOCK);
+                .add(ModBlocks.REDSTONE_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.AMETHYST_ORE_BLOCK)
+                .add(ModBlocks.CUPRIUM_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.CUPRIUM_ORE_BLOCK)
+                .add(ModBlocks.DEEPSLATE_AMETHYST_ORE)
+                .add(ModBlocks.DEEPSLATE_FERRIUM_ORE)
+                .add(ModBlocks.DEEPSLATE_CUPRIUM_ORE)
+                .add(ModBlocks.DEEPSLATE_ORIGINIUM_ORE)
+                .add(ModBlocks.FERRIUM_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.FERRIUM_ORE_BLOCK)
+                .add(ModBlocks.ORIGINIUM_ORE_BLOCK);
 
         getOrCreateTagBuilder(BlockTags.NEEDS_DIAMOND_TOOL)
                 .add(ModBlocks.AMETHYST_MINERAL_VEIN_BLOCK)
@@ -37,6 +47,20 @@ public class ModBlockTagsProvider extends FabricTagProvider.BlockTagProvider {
                 .add(ModBlocks.IRON_MINERAL_VEIN_BLOCK)
                 .add(ModBlocks.LAPIS_MINERAL_VEIN_BLOCK)
                 .add(ModBlocks.ORIGINIUM_MINERAL_VEIN_BLOCK)
-                .add(ModBlocks.REDSTONE_MINERAL_VEIN_BLOCK);
+                .add(ModBlocks.REDSTONE_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.CUPRIUM_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.FERRIUM_MINERAL_VEIN_BLOCK);
+
+        getOrCreateTagBuilder(BlockTags.NEEDS_STONE_TOOL)
+                .add(ModBlocks.ORIGINIUM_ORE_BLOCK)
+                .add(ModBlocks.DEEPSLATE_ORIGINIUM_ORE);
+
+        getOrCreateTagBuilder(BlockTags.NEEDS_IRON_TOOL)
+                .add(ModBlocks.FERRIUM_ORE_BLOCK)
+                .add(ModBlocks.DEEPSLATE_FERRIUM_ORE)
+                .add(ModBlocks.AMETHYST_ORE_BLOCK)
+                .add(ModBlocks.DEEPSLATE_AMETHYST_ORE)
+                .add(ModBlocks.CUPRIUM_ORE_BLOCK)
+                .add(ModBlocks.DEEPSLATE_CUPRIUM_ORE);
     }
 }